JS: Remove an unnecessary use of Type
If a JSON-like object is returned, HTML injection is not possible because the content-type will be set to application/json. There was an attempt to avoid FPs by checking the TypeScript-inferred return type.

However, in these cases it is mostly harmless to use the return value as a sink, because an object as a whole will very rarely be tainted.